Add dev and source code link Add Linter Add SCSS and SASS Refactored code Changed folder structure
