+++ docs/asset-manifest.json
@@ -1,23 +1,23 @@
 {
   "files": {
-    "main.css": "/static/css/main.8ab6f918.chunk.css",
-    "main.js": "/static/js/main.bf422369.chunk.js",
-    "main.js.map": "/static/js/main.bf422369.chunk.js.map",
+    "main.css": "/static/css/main.4367f119.chunk.css",
+    "main.js": "/static/js/main.ad93a0bd.chunk.js",
+    "main.js.map": "/static/js/main.ad93a0bd.chunk.js.map",
     "runtime-main.js": "/static/js/runtime-main.d8a73986.js",
     "runtime-main.js.map": "/static/js/runtime-main.d8a73986.js.map",
     "static/js/2.fc02ac24.chunk.js": "/static/js/2.fc02ac24.chunk.js",
     "static/js/2.fc02ac24.chunk.js.map": "/static/js/2.fc02ac24.chunk.js.map",
     "index.html": "/index.html",
-    "precache-manifest.cc4fd6e9ac6a32ef5ec68bafaa226ee1.js": "/precache-manifest.cc4fd6e9ac6a32ef5ec68bafaa226ee1.js",
+    "precache-manifest.fe02c12eb5c5e26660e9a10a417e7959.js": "/precache-manifest.fe02c12eb5c5e26660e9a10a417e7959.js",
     "service-worker.js": "/service-worker.js",
-    "static/css/main.8ab6f918.chunk.css.map": "/static/css/main.8ab6f918.chunk.css.map",
+    "static/css/main.4367f119.chunk.css.map": "/static/css/main.4367f119.chunk.css.map",
     "static/js/2.fc02ac24.chunk.js.LICENSE": "/static/js/2.fc02ac24.chunk.js.LICENSE",
     "static/media/logo.svg": "/static/media/logo.ee7cd8ed.svg"
   },
   "entrypoints": [
     "static/js/runtime-main.d8a73986.js",
     "static/js/2.fc02ac24.chunk.js",
-    "static/css/main.8ab6f918.chunk.css",
-    "static/js/main.bf422369.chunk.js"
+    "static/css/main.4367f119.chunk.css",
+    "static/js/main.ad93a0bd.chunk.js"
   ]
 }
