+++ docs/static/js/main.ad93a0bd.chunk.js
@@ -1,0 +1,2 @@
+(this.webpackJsonpdmt2=this.webpackJsonpdmt2||[]).push([[0],[,,,function(e,n,t){e.exports=t.p+"static/media/logo.ee7cd8ed.svg"},function(e,n,t){e.exports=t(11)},,,,,function(e,n,t){},function(e,n,t){},function(e,n,t){"use strict";t.r(n);var o=t(0),a=t.n(o),r=t(2),i=t.n(r),c=(t(9),t(3)),l=t.n(c),s=function(){return a.a.createElement(a.a.Fragment,null,a.a.createElement("h1",null,"Oh shit, here we go again"),a.a.createElement("img",{src:l.a,alt:""}))};t(10),Boolean("localhost"===window.location.hostname||"[::1]"===window.location.hostname||window.location.hostname.match(/^127(?:\.(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)){3}$/));i.a.render(a.a.createElement((function(){return a.a.createElement(s,null)}),null),document.getElementById("root")),"serviceWorker"in navigator&&navigator.serviceWorker.ready.then((function(e){e.unregister()}))}],[[4,1,2]]]);
+//# sourceMappingURL=main.ad93a0bd.chunk.js.map
