+++ src/app/App.test.tsx
@@ -1,5 +1,6 @@
-import React from 'react';
 import { render } from '@testing-library/react';
+import './App.css';
+import React from 'react';
 import App from './App';
 
 test('renders learn react link', () => {
