+++ src/app/App.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import './App.scss';
+import logo from '../img/logo.svg';
+
+const App: React.FC = () => {
+  return (
+    <>
+      <h1>Oh shit, here we go again</h1>
+      <img src={ logo } alt="" />
+    </>
+  );
+};
+
+export default App;
