+++ tslint.json
@@ -1,0 +1,90 @@
+{
+  "extends": [
+    "tslint:recommended",
+    "tslint-config-prettier",
+    "tslint-plugin-prettier"
+  ],
+  "rules": {
+    // Formatting rules
+    "prettier": false,
+    "array-type": false,
+    // TODO: Enable stricter options for these
+    "comment-format": {
+      "options": [
+        "check-space"
+      ]
+    },
+    "member-access": true,
+    "only-arrow-functions": {
+      "options": [
+        "allow-declarations",
+        "allow-named-functions"
+      ]
+    },
+    "curly": false,
+    // TODO: Enable these
+    "completed-docs": false,
+    "no-any": false,
+    "no-magic-numbers": false,
+    "no-non-null-assertion": false,
+    "no-null-keyword": false,
+    "no-require-imports": false,
+    "no-unbound-method": false,
+    "no-unnecessary-qualifier": false,
+    "no-use-before-declare": false,
+    "no-void-expression": false,
+    "object-literal-sort-keys": false,
+    "prefer-function-over-method": false,
+    "strict-type-predicates": false,
+    "triple-equals": {
+      "options": [
+        "allow-undefined-check"
+      ]
+    },
+    // Misc
+    "ban": {
+      "options": [
+        [
+          "describe",
+          "only"
+        ],
+        [
+          "it",
+          "only"
+        ]
+      ]
+    },
+    "interface-name": false,
+    "max-classes-per-file": false,
+    "member-ordering": {
+      "options": {
+        "order": "statics-first"
+      }
+    },
+    "quotemark": [
+      true,
+      "single"
+    ],
+    "no-console": {
+      "options": [
+        "log"
+      ]
+    },
+    "no-switch-case-fall-through": true,
+    "strict-boolean-expressions": {
+      "options": [
+        "allow-boolean-or-undefined"
+      ]
+    },
+    "switch-default": false,
+    "variable-name": {
+      "options": [
+        "ban-keywords",
+        "check-format",
+        "allow-leading-underscore",
+        "allow-pascal-case"
+      ]
+    },
+    "linebreak-style": false
+  }
+}
